requirements.txt
fastapi
pandas
yfinance
python-dateutil
uvicorn
gunicorn
jinja2
python-multipart
